+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Issue #1317: encode the share link.
 - MouseHover: Prevent duplicate hover info for WebGL polygon and line layers.
 - Issue #1527: statisticDashboard - restored WFS support for statistic dashboard and added default color palettes.
+- Issue #1531: selectFeatures - fix link display in selectFeatures tool.
 - WFS-cluster: Fix a recursive style loop that triggered a ‘Maximum call stack size exceeded’ error.
 - Issue #1168: Fix autoRefresh for LayerGroups.
 - MapMarker: Fix marker visibility after ShareView.